Add script to generate a dataset
The script extracts skeletons from all videos in a directory and
generates annotations for each one in CSV format as well as the output
video with the predicted skeletons.

CAVEAT! Path to the dataset is HARDCODED!

The CSV contains all the keypoints of the person in the video,
in ASLLVD videos there is a single person.
Each keypoint is represented by 3 values: x1, y1, c1.
- x1, y1: Image coordinates
- c1: confidence of the keypoint

CAVEAT! CSV and videos output path are HARDCODED!
CAVEAT! It is HARDCODED to annotate only the first person.